# swagger-documentation.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Registration Successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/verify-email/{token}":{"post":{"operationId":"AuthController_confirmEmail","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/resend-email-verification-code":{"post":{"operationId":"AuthController_resendEmailVerificationCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/resend-forgot-password":{"post":{"operationId":"AuthController_resendForgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/reset-password":{"patch":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/change-password":{"patch":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/logout":{"get":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/users":{"get":{"operationId":"UsersController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Users retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}},"tags":["Users"]}},"/api/v1/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"user updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"user updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"User deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Users"]}},"/api/v1/users/me":{"get":{"operationId":"UsersController_findMe","parameters":[],"responses":{"200":{"description":"user retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["Users"]}},"/api/v1/users/upload":{"put":{"operationId":"UsersController_UploadUserPhoto","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"photo":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"user photo uploaded successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["Users"]}},"/api/v1/rooms":{"post":{"operationId":"RoomsController_create","parameters":[],"requestBody":{"required":true,"description":"Create a new room","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomDto"}}}},"responses":{"201":{"description":"room created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}},"tags":["Rooms"]},"get":{"operationId":"RoomsController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Rooms retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomDto"}}}}}},"tags":["Rooms"]}},"/api/v1/rooms/{id}":{"get":{"operationId":"RoomsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Room retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}},"tags":["Rooms"]},"patch":{"operationId":"RoomsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoomsDto"}}}},"responses":{"200":{"description":"Room Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}},"tags":["Rooms"]},"delete":{"operationId":"RoomsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Room deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Rooms"]}}},"info":{"title":"Colkie Chat App","description":"Colkie Senior Backend Developer Task (chat-app)","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8989"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"password":{"type":"string"},"passwordConfirm":{"type":"string"}},"required":["name","email","phoneNumber","password","passwordConfirm"]},"User":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"reachme@amitavroy.com"},"password":{"type":"string","description":"Password in plain text","example":"Password@123"}},"required":["email","password"]},"EmailDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"reachme@amitavroy.com"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"password":{"type":"string","description":"Password in plain text","example":"Password@123"},"passwordConfirm":{"type":"string","description":"Password in plain text","example":"Password@123"},"resetToken":{"type":"string","description":"Reset Code sent to Emailt","example":"4972fe"}},"required":["password","passwordConfirm","resetToken"]},"ChangePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string","description":"Password in plain text","example":"Password@123"},"newPassword":{"type":"string","description":"Password in plain text","example":"Password@123"},"newPasswordConfirm":{"type":"string","description":"Password in plain text","example":"Password@123"}},"required":["oldPassword","newPassword","newPasswordConfirm"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"photo":{"type":"string"},"isEmailVerified":{"type":"boolean"},"role":{"type":"string","enum":["user","manager","admin"]}},"required":["id","name","email","phoneNumber","photo","isEmailVerified","role"]},"UpdateUserDto":{"type":"object","properties":{}},"StandardResponse":{"type":"object","properties":{"status":{"type":"string"},"code":{"type":"number"},"message":{"type":"string"},"data":{"type":"object"},"meta":{"type":"object"}},"required":["status","code","message","data","meta"]},"CreateRoomDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]},"RoomDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"users":{"type":"array","items":{"type":"string"}},"messages":{"type":"array","items":{"type":"string"}}},"required":["title","description","users","messages"]},"UpdateRoomsDto":{"type":"object","properties":{}}}}}